requirements.txt
attrs==25.1.0
certifi==2025.1.31
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
fiona==1.10.1
geopandas==1.0.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
psycopg2==2.9.10
pyogrio==0.10.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
shapely==2.0.7
six==1.17.0
tzdata==2025.1
